continuation algo

+++ src/input_creation_and_iterative_methods.h
@@ -5,7 +5,7 @@
 #include <math.h>
 #include <assert.h>
 
-int normalize_vector(gsl_vector *vector_to_normalize);
+double normalize_vector(gsl_vector *vector_to_normalize);
 double scalar_product(gsl_vector *vector1, gsl_vector *vector2);
 gsl_vector* matrix_vector_product(gsl_matrix *matrix, gsl_vector *vector);
 void print_vector_contents(gsl_vector *vector);
